+++ src-docs/src/views/filter_group/filter_group.js
@@ -5,14 +5,10 @@ import React, {
 
 import {
   EuiPopover,
-  EuiPopoverTitle,
-  EuiFieldSearch,
-  EuiFilterSelectItem,
-  EuiLoadingChart,
-  EuiSpacer,
-  EuiIcon,
   EuiFilterGroup,
   EuiFilterButton,
+  EuiIcon,
+  EuiSpacer,
 } from '../../../../src/components';
 
 export default class extends Component {
@@ -61,51 +57,28 @@ export default class extends Component {
 
   render() {
 
-    const items = [
-      { name: 'Johann Sebastian Bach', checked: 'on' },
-      { name: 'Wolfgang Amadeus Mozart', checked: 'on' },
-      { name: 'Antonín Dvořák', checked: 'off' },
-      { name: 'Dmitri Shostakovich' },
-      { name: 'Felix Mendelssohn-Bartholdy' },
-      { name: 'Franz Liszt' },
-      { name: 'Franz Schubert' },
-      { name: 'Frédéric Chopin' },
-      { name: 'Georg Friedrich Händel' },
-      { name: 'Giuseppe Verdi' },
-      { name: 'Gustav Mahler' },
-      { name: 'Igor Stravinsky' },
-      { name: 'Johannes Brahms' },
-      { name: 'Joseph Haydn' },
-      { name: 'Ludwig van Beethoven' },
-      { name: 'Piotr Illitch Tchaïkovsky' },
-      { name: 'Robert Schumann' },
-      { name: 'Sergej S. Prokofiew' },
-      { name: 'Wolfgang Amadeus Mozart' },
-    ];
-
     const button = (
       <EuiFilterButton
         iconType="arrowDown"
         onClick={this.onButtonClick.bind(this)}
         isSelected={this.state.isPopoverOpen}
-        numFilters={items.length}
+        numFilters={12}
         hasActiveFilters={true}
         numActiveFilters={2}
-        grow={true}
       >
         Composers
       </EuiFilterButton>
     );
 
     return (
-      <EuiFilterGroup>
-        <EuiFilterButton hasActiveFilters={this.state.isFilterOn} onClick={this.toggleFilter}>
+      <EuiFilterGroup fullWidth={true}>
+        <EuiFilterButton grow={false} hasActiveFilters={this.state.isFilterOn} onClick={this.toggleFilter}>
           Filter
         </EuiFilterButton>
-        <EuiFilterButton noDivider hasActiveFilters={this.state.isOnFilterOn} onClick={this.toggleOnFilter}>
+        <EuiFilterButton withNext grow={false} hasActiveFilters={this.state.isOnFilterOn} onClick={this.toggleOnFilter}>
           On
         </EuiFilterButton>
-        <EuiFilterButton hasActiveFilters={this.state.isOffFilterOn} onClick={this.toggleOffFilter}>
+        <EuiFilterButton grow={false} hasActiveFilters={this.state.isOffFilterOn} onClick={this.toggleOffFilter}>
           Off
         </EuiFilterButton>
         <EuiPopover
@@ -117,40 +90,17 @@ export default class extends Component {
           panelPaddingSize="none"
           withTitle
         >
-          <EuiPopoverTitle>
-            <EuiFieldSearch />
-          </EuiPopoverTitle>
-          <div className="euiFilterSelect__items">
-            {items.map((item, index) => (
-              <EuiFilterSelectItem
-                checked={item.checked}
-                key={index}
-              >
-                {item.name}
-              </EuiFilterSelectItem>
-            ))}
-            {/*
-              Use when loading items initially
-            */}
-            <div className="euiFilterSelect__note">
-              <div className="euiFilterSelect__noteContent">
-                <EuiLoadingChart size="m" />
-                <EuiSpacer size="xs" />
-                <p>Loading filters</p>
-              </div>
-            </div>
-            {/*
-              Use when no results are returned
-            */}
-            <div className="euiFilterSelect__note">
-              <div className="euiFilterSelect__noteContent">
-                <EuiIcon type="minusInCircle" />
-                <EuiSpacer size="xs" />
-                <p>No filters found</p>
-              </div>
+          <div className="euiFilterSelect__note">
+            <div className="euiFilterSelect__noteContent">
+              <EuiIcon type="minusInCircle" />
+              <EuiSpacer size="xs" />
+              <p>No filters found</p>
             </div>
           </div>
         </EuiPopover>
+        <EuiFilterButton numFilters={12} hasActiveFilters={this.state.isFilterOn} onClick={this.toggleFilter}>
+          Filter with a very long name
+        </EuiFilterButton>
       </EuiFilterGroup>
     );
   }