+++ src/components/filter_group/filter_group.test.tsx
@@ -1,0 +1,21 @@
+import React from 'react';
+import { render } from 'enzyme';
+import { requiredProps } from '../../test';
+
+import { EuiFilterGroup } from './filter_group';
+
+describe('EuiFilterGroup', () => {
+  test('is rendered', () => {
+    const component = render(<EuiFilterGroup {...requiredProps} />);
+
+    expect(component).toMatchSnapshot();
+  });
+
+  describe('props', () => {
+    test('fullWidth is rendered', () => {
+      const component = render(<EuiFilterGroup fullWidth />);
+
+      expect(component).toMatchSnapshot();
+    });
+  });
+});