+++ src/components/search_bar/filters/field_value_toggle_group_filter.js
@@ -67,7 +67,7 @@ export class FieldValueToggleGroupFilter extends Component {
           key={key}
           onClick={onClick}
           hasActiveFilters={active}
-          noDivider={!isLastItem}
+          withNext={!isLastItem}
         >
           {name}
         </EuiFilterButton>